cram/codecs/rans_4x8/encode: Reduce frequency size
Frequencies aren't greater than 0x0fff.

+++ noodles-cram/src/codecs/rans_4x8/encode/order_0.rs
@@ -44,7 +44,7 @@ pub fn encode(src: &[u8]) -> io::Result<Vec<u8>> {
     Ok(dst)
 }
 
-pub fn write_frequencies<W>(writer: &mut W, frequencies: &[u32]) -> io::Result<()>
+pub fn write_frequencies<W>(writer: &mut W, frequencies: &[u16]) -> io::Result<()>
 where
     W: Write,
 {
